refactor: relocate the per-site viewmodelbuilders into the sites (#1619)
