+++ site-config/brc-analytics/local/entity/organism/organismMainColumn.ts
@@ -1,4 +1,4 @@
-import * as V from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
+import * as V from "@brc/viewModelBuilders/viewModelBuilders";
 import { Main as OrganismViewMain } from "@brc/views/OrganismView/components/Main/main";
 import { BackPageContentSingleColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import { type ComponentsConfig } from "@databiosphere/findable-ui/lib/config/entities";
