+++ site-config/brc-analytics/local/entity/priorityPathogen/priorityPathogenMainColumn.ts
@@ -1,5 +1,5 @@
-import * as V from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
 import { type Outbreak } from "@brc/apis/outbreak";
+import * as V from "@brc/viewModelBuilders/viewModelBuilders";
 import { ResourcesSection } from "@brc/views/PriorityPathogenView/components/ResourcesSection/resourcesSection";
 import { MDXSection } from "@brc/views/PriorityPathogenView/ui/Section/MDXSection/mdxSection";
 import { BackPageContentMainColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
