+++ site-config/brc-analytics/local/entity/priorityPathogen/priorityPathogenSideColumn.ts
@@ -1,5 +1,5 @@
-import * as V from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
 import { type Outbreak } from "@brc/apis/outbreak";
+import * as V from "@brc/viewModelBuilders/viewModelBuilders";
 import { KeyValueSection } from "@brc/views/PriorityPathogenView/ui/Section/KeyValueSection/keyValueSection";
 import { Sections } from "@brc/views/PriorityPathogenView/ui/Sections/sections";
 import { BackPageContentSideColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
