+++ site-config/brc-analytics/local/index/genomeEntityConfig.ts
@@ -1,6 +1,6 @@
-import * as V from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
 import { type BRCDataCatalogGenome } from "@brc/apis/assembly";
 import { getGenomeId, getGenomeTitle } from "@brc/apis/utils";
+import * as V from "@brc/viewModelBuilders/viewModelBuilders";
 import { BasicCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/BasicCell/basicCell";
 import { ChipCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/ChipCell/chipCell";
 import { NTagCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/NTagCell/nTagCell";
