+++ site-config/ga2/local/entity/organism/organismMainColumn.ts
@@ -1,10 +1,10 @@
-import * as V from "@/viewModelBuilders/catalog/ga2/viewModelBuilders";
 import { BackPageContentSingleColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import {
   type ComponentConfig,
   type ComponentsConfig,
 } from "@databiosphere/findable-ui/lib/config/entities";
 import { type GA2OrganismEntity } from "@ga2/apis/organism";
+import * as V from "@ga2/viewModelBuilders/viewModelBuilders";
 import { Main as OrganismViewMain } from "@ga2/views/OrganismView/components/Main/main";
 
 export const organismMainColumn: ComponentsConfig = [
