+++ site-config/ga2/local/entity/organism/organismTop.ts
@@ -1,10 +1,10 @@
-import * as V from "@/viewModelBuilders/catalog/ga2/viewModelBuilders";
 import { BackPageHero } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/components/BackPageHero/backPageHero";
 import {
   type ComponentConfig,
   type ComponentsConfig,
 } from "@databiosphere/findable-ui/lib/config/entities";
 import { type GA2OrganismEntity } from "@ga2/apis/organism";
+import * as V from "@ga2/viewModelBuilders/viewModelBuilders";
 
 export const organismTop: ComponentsConfig = [
   {
