+++ site-config/ga2/local/index/genome/columnDefs.ts
@@ -23,8 +23,7 @@ import {
   buildTaxonomicLevelOrder,
   buildTaxonomicLevelPhylum,
   buildTaxonomyId,
-} from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
-import * as V from "@/viewModelBuilders/catalog/ga2/viewModelBuilders";
+} from "@brc/viewModelBuilders/viewModelBuilders";
 import { BasicCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/BasicCell/basicCell";
 import { ChipCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/ChipCell/chipCell";
 import { NTagCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/NTagCell/nTagCell";
@@ -34,6 +33,7 @@ import {
 } from "@databiosphere/findable-ui/lib/config/entities";
 import { type GA2AssemblyEntity } from "@ga2/apis/assembly";
 import { OrganismAvatar } from "@ga2/components/OrganismAvatar/organismAvatar";
+import * as V from "@ga2/viewModelBuilders/viewModelBuilders";
 import { AnalyzeGenome } from "@repo/shared/components/Table/components/TableCell/components/AnalyzeGenome/analyzeGenome";
 import { LevelCell } from "@repo/shared/components/Table/components/TableCell/components/LevelCell/levelCell";
 import { SpeciesCell } from "@repo/shared/components/Table/components/TableCell/components/SpeciesCell/speciesCell";
