+++ site-config/ga2/local/index/organism/columnDefs.ts
@@ -9,8 +9,7 @@ import {
   buildTaxonomicLevelKingdom,
   buildTaxonomicLevelOrder,
   buildTaxonomicLevelPhylum,
-} from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
-import * as V from "@/viewModelBuilders/catalog/ga2/viewModelBuilders";
+} from "@brc/viewModelBuilders/viewModelBuilders";
 import { Link } from "@databiosphere/findable-ui/lib/components/Links/components/Link/link";
 import { BasicCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/BasicCell/basicCell";
 import { NTagCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/NTagCell/nTagCell";
@@ -20,6 +19,7 @@ import {
 } from "@databiosphere/findable-ui/lib/config/entities";
 import { type GA2OrganismEntity } from "@ga2/apis/organism";
 import { OrganismAvatar } from "@ga2/components/OrganismAvatar/organismAvatar";
+import * as V from "@ga2/viewModelBuilders/viewModelBuilders";
 import {
   GA2_CATEGORY_KEY,
   GA2_CATEGORY_LABEL,
