+++ sites/brc-analytics/views/EntityView/assembly/components/Side/brc/side.tsx
@@ -2,7 +2,7 @@ import {
   buildAssemblyDetails,
   buildAssemblyResources,
   buildOrganismDetails,
-} from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
+} from "@brc/viewModelBuilders/viewModelBuilders";
 import { BackPageContentSideColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import { AssemblyFavoriteButton } from "@repo/shared/components/Favorites/components/AssemblyFavoriteButton/assemblyFavoriteButton";
 import { AnalysisPortals } from "@repo/shared/views/EntityView/assembly/components/Side/AnalysisPortals/analysisPortals";
