+++ tests/viewModelBuilders/species.test.ts
@@ -1,13 +1,13 @@
+import type { BRCDataCatalogGenome } from "@brc/apis/assembly";
 import {
   buildGenomeSpecies,
   buildOrganismGenomeSpecies,
-} from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
+} from "@brc/viewModelBuilders/viewModelBuilders";
+import type { GA2AssemblyEntity } from "@ga2/apis/assembly";
 import {
   buildAssemblySpecies,
   buildOrganismAssemblySpecies,
-} from "@/viewModelBuilders/catalog/ga2/viewModelBuilders";
-import type { BRCDataCatalogGenome } from "@brc/apis/assembly";
-import type { GA2AssemblyEntity } from "@ga2/apis/assembly";
+} from "@ga2/viewModelBuilders/viewModelBuilders";
 
 describe("buildGenomeSpecies", () => {
   test("surfaces species, taxonomy id and all populated minor fields", () => {
